pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bonboneka"
version = "0.1.0"
description = "Bundle HTML/CSS/JS into an Android WebView app"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Brickboss", email = "info@example.com"}
]
keywords = ["android", "webview", "app", "builder", "html", "css", "javascript"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.scripts]
bomk = "bomk.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/bonboneka"
"Bug Reports" = "https://github.com/yourusername/bonboneka/issues"
"Source Code" = "https://github.com/yourusername/bonboneka"
